Commit validation work
Commits validation checks for the server controls that must be specified by the user before processing an order, or updating the totalCostLabel. Adds a Success.aspx web form, which the user is redirected to after placing a successful order. Bug: totalCostLabel updates when Pepperoni is checked/unchecked, but checking/unchecking other toppings does not update totalCostLabel.